requirements.txt
# streamlit
# sentence-transformers
# faiss-cpu
# PyPDF2>=3.0.0
# numpy>=1.24.0
# scikit-learn>=1.2.0
# openai>=0.27.0
# tqdm>=4.65.0
# torch
# torchvision
# torchaudio




streamlit
sentence-transformers
faiss-cpu
PyPDF2>=3.0.0
numpy>=1.24.0
scikit-learn>=1.2.0
tqdm>=4.65.0
torch
torchvision
torchaudio
groq
